Add support for serializing by the user
The user may opt to register Map for writing and reading chunks of
map data.

For writing, the user provides a function that takes an Enumerable<T> of
their items and a co-ordinate 3-Tuple, they can then write that data as
they please.

For reading, the user provides a function that takes a 3-Tuple
coordinate, they can then return a Enumerable of the same length as
the dimensions of the map chunks (height*width*depth) from any source
they want to use.

This makes it flexible for the user. They can do something like
writing to a file and reading back again, i.e. persistence. Or the user
could discard all unloaded chunks and generate (read) chunks from
a generator function.

This commit adds most of the basic functionality to support #2 - the
rest of the work here will be for #3 - namely, unloading chunks.