Adds check for creation in GetChunk
If createIfNull is set, then we give the user a chance to create a new
chunk, the intention being that a normal GetChunk is issued when walking
the collection and the user wants to fill in blank areas with data.

But when unloading areas, we don't want to create a blank or generated
chunk to then immediately discard it, it's a waste of time and may cause
hard to trace errors in the users world generation methods.